Finish separating inline test blocks outside the curated hotspots (#687)
Seven production files over the 1,000-line review budget still carried an
inline `#[cfg(test)] mod` block. They were every remaining such file outside
the eight curated hotspot aggregates:

| file | blocks | before | after |
| --- | --- | --- | --- |
| wow-packet/packets/quest.rs | 6 | 1,443 | 978 |
| bnet-server/rest/handlers.rs | 1 | 1,304 | 1,057 |
| wow-database/battle_pet_account_adapter.rs | 1 | 1,271 | 1,227 |
| wow-data/item_stats.rs | 1 | 1,268 | 1,050 |
| wow-packet/packets/query.rs | 2 | 1,216 | 1,027 |
| wow-data/player_condition.rs | 1 | 1,168 | 1,037 |
| wow-database/transaction.rs | 1 | 1,104 | 971 |

#685 moved only the block named `tests`, which is why the five differently
named blocks in `packets/quest.rs` and both blocks in `packets/query.rs`
survived it. A block named `tests` becomes `<stem>/tests/mod.rs`; a
differently named block becomes `<stem>/<name>.rs`, mounted under its own
name, so `quest::quest_push_result_tests::…` keeps its path instead of
becoming `quest::tests::…`. `#[path]` mounting leaves the module hierarchy
unchanged, so each moved block's `use super::*;` still resolves to its owner.

The files inside the curated hotspots keep their inline blocks -
`world-server/creature_loaded_grid.rs`, `spell_acquisition_loader.rs`,
`area_trigger_loaded_grid.rs`, `spell_core_db2_hotfix.rs` and
`session/directory.rs`. Each hotspot's test lines are an independent
non-growth ceiling, so relocating tests within the aggregate the ratchet
measures only adds mount lines. Their exit is the semantic extraction #584
owns, not a physical move.

Physical ratchet cohesion reviews drop from 94 to 92; none of the seven
files carries a legacy ceiling row, so no ceiling moved.

Test name multiset identical across the workspace at 8,579 entries, and test
counts identical in every affected crate - wow-packet 728, wow-data 724,
wow-database 358, bnet-server 87. No production line, test body, assertion
or fixture changed. Architecture check, self-test, physical ratchet, hotspot
ratchet and the syntax-only session ownership check pass.

Closes #687
