+++ crates/wow-packet/src/packets/query.rs
@@ -1019,197 +1019,9 @@ impl ServerPacket for RealmQueryResponse {
 }
 
 #[cfg(test)]
-mod quest_poi_tests {
-    use super::*;
-    use num_traits::ToPrimitive;
-
-    fn client_payload(bytes: &[u8]) -> WorldPacket {
-        let mut data = Vec::from(ClientOpcodes::QuestPoiQuery.to_u16().unwrap().to_le_bytes());
-        data.extend_from_slice(bytes);
-        let mut pkt = WorldPacket::from_bytes(&data);
-        pkt.skip_opcode();
-        pkt
-    }
-
-    #[test]
-    fn quest_poi_query_reads_fixed_quest_log_array_like_cpp() {
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&2i32.to_le_bytes());
-        for i in 0..QUEST_POI_QUERY_MISSING_QUEST_POIS_LIKE_CPP {
-            payload.extend_from_slice(&(1000 + i as i32).to_le_bytes());
-        }
-
-        let mut pkt = client_payload(&payload);
-        let parsed = QuestPoiQuery::read(&mut pkt).unwrap();
-
-        assert_eq!(parsed.missing_quest_count, 2);
-        assert_eq!(parsed.missing_quest_pois[0], 1000);
-        assert_eq!(parsed.missing_quest_pois[1], 1001);
-        assert_eq!(parsed.missing_quest_pois[24], 1024);
-    }
-
-    #[test]
-    fn quest_poi_query_reads_real_client_len_106_shape_like_cpp() {
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&25i32.to_le_bytes());
-        for i in 0..25 {
-            payload.extend_from_slice(&(2000i32 + i).to_le_bytes());
-        }
-
-        let mut pkt = client_payload(&payload);
-        let parsed = QuestPoiQuery::read(&mut pkt).unwrap();
-
-        assert_eq!(parsed.missing_quest_count, 25);
-        assert_eq!(parsed.missing_quest_pois[0], 2000);
-        assert_eq!(parsed.missing_quest_pois[24], 2024);
-    }
-
-    #[test]
-    fn quest_poi_query_rejects_short_fixed_quest_log_array_like_cpp() {
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&1i32.to_le_bytes());
-        payload.extend_from_slice(&123i32.to_le_bytes());
-
-        let mut pkt = client_payload(&payload);
-
-        assert!(matches!(
-            QuestPoiQuery::read(&mut pkt),
-            Err(PacketError::ReadPastEnd { .. })
-        ));
-    }
-
-    #[test]
-    fn quest_poi_query_response_serializes_cpp_shape() {
-        let response = QuestPoiQueryResponse {
-            quest_poi_data_stats: vec![QuestPoiData {
-                quest_id: 77,
-                blobs: vec![QuestPoiBlobData {
-                    blob_index: 1,
-                    objective_index: -1,
-                    quest_objective_id: 2,
-                    quest_object_id: 3,
-                    map_id: 571,
-                    ui_map_id: 486,
-                    priority: 4,
-                    flags: 5,
-                    world_effect_id: 6,
-                    player_condition_id: 7,
-                    navigation_player_condition_id: 8,
-                    spawn_tracking_id: 9,
-                    points: vec![QuestPoiBlobPoint {
-                        x: 10,
-                        y: -11,
-                        z: 12,
-                    }],
-                    always_allow_merging_blobs: true,
-                }],
-            }],
-        };
-
-        let bytes = response.to_bytes();
-        let mut expected = Vec::from(
-            ServerOpcodes::QuestPoiQueryResponse
-                .to_u16()
-                .unwrap()
-                .to_le_bytes(),
-        );
-        expected.extend_from_slice(&1i32.to_le_bytes());
-        expected.extend_from_slice(&1i32.to_le_bytes());
-        expected.extend_from_slice(&77i32.to_le_bytes());
-        expected.extend_from_slice(&1i32.to_le_bytes());
-        for value in [1, -1, 2, 3, 571, 486, 4, 5, 6, 7, 8, 9, 1] {
-            expected.extend_from_slice(&i32::to_le_bytes(value));
-        }
-        expected.extend_from_slice(&10i16.to_le_bytes());
-        expected.extend_from_slice(&(-11i16).to_le_bytes());
-        expected.extend_from_slice(&12i16.to_le_bytes());
-        expected.push(0x80);
-
-        assert_eq!(bytes, expected);
-    }
-}
+#[path = "query/quest_poi_tests.rs"]
+mod quest_poi_tests;
 
 #[cfg(test)]
-mod query_quest_completion_tests {
-    use super::*;
-    use num_traits::ToPrimitive;
-
-    fn client_payload(bytes: &[u8]) -> WorldPacket {
-        let mut data = Vec::from(
-            ClientOpcodes::QueryQuestCompletionNpcs
-                .to_u16()
-                .unwrap()
-                .to_le_bytes(),
-        );
-        data.extend_from_slice(bytes);
-        let mut pkt = WorldPacket::from_bytes(&data);
-        pkt.skip_opcode();
-        pkt
-    }
-
-    #[test]
-    fn query_quest_completion_parse_valid_ids() {
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&2u32.to_le_bytes());
-        payload.extend_from_slice(&123i32.to_le_bytes());
-        payload.extend_from_slice(&(-45i32).to_le_bytes());
-
-        let mut pkt = client_payload(&payload);
-        let parsed = QueryQuestCompletionNpcs::read(&mut pkt).unwrap();
-
-        assert_eq!(parsed.quest_ids, vec![123, -45]);
-    }
-
-    #[test]
-    fn query_quest_completion_rejects_short_payload() {
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&1u32.to_le_bytes());
-        payload.extend_from_slice(&[0x34, 0x12]);
-
-        let mut pkt = client_payload(&payload);
-
-        assert!(matches!(
-            QueryQuestCompletionNpcs::read(&mut pkt),
-            Err(PacketError::ReadPastEnd { .. })
-        ));
-    }
-
-    #[test]
-    fn query_quest_completion_rejects_over_cap() {
-        let mut payload = Vec::new();
-        payload.extend_from_slice(&101u32.to_le_bytes());
-
-        let mut pkt = client_payload(&payload);
-
-        assert!(matches!(
-            QueryQuestCompletionNpcs::read(&mut pkt),
-            Err(PacketError::TooLarge { size: 101 })
-        ));
-    }
-
-    #[test]
-    fn query_quest_completion_response_serializes_creature_and_masked_go() {
-        let response = QuestCompletionNpcResponse {
-            quests: vec![QuestCompletionNpc {
-                quest_id: 77,
-                npcs: vec![1234, 0x8000_5678u32 as i32],
-            }],
-        };
-
-        let bytes = response.to_bytes();
-
-        let mut expected = Vec::from(
-            ServerOpcodes::QuestCompletionNpcResponse
-                .to_u16()
-                .unwrap()
-                .to_le_bytes(),
-        );
-        expected.extend_from_slice(&1u32.to_le_bytes());
-        expected.extend_from_slice(&77i32.to_le_bytes());
-        expected.extend_from_slice(&2u32.to_le_bytes());
-        expected.extend_from_slice(&1234i32.to_le_bytes());
-        expected.extend_from_slice(&0x8000_5678u32.to_le_bytes());
-
-        assert_eq!(bytes, expected);
-    }
-}
+#[path = "query/query_quest_completion_tests.rs"]
+mod query_quest_completion_tests;
